feat: export upload utils (#1252)
