+++ cli/upload/index.ts
@@ -1,0 +1,2 @@
+export * from './uploadSourcemaps';
+export * from './uploadSoFiles';
