+++ rollup.config.js
@@ -5,19 +5,30 @@ import typescript from '@rollup/plugin-typescript';
 import cleanup from 'rollup-plugin-cleanup';
 import shebang from 'rollup-plugin-preserve-shebang';
 
+const commonPlugins = [
+  shebang(),
+  json(),
+  nodeResolve({ preferBuiltins: true }),
+  commonjs(),
+  cleanup(),
+];
+
 /** @type import('rollup').RollupOptions */
-export default {
-  input: ['cli/index.ts'],
-  output: {
-    dir: 'bin',
-    format: 'cjs',
+export default [
+  {
+    input: ['cli/index.ts'],
+    output: {
+      dir: 'bin',
+      format: 'cjs',
+    },
+    plugins: [...commonPlugins, typescript({ tsconfig: './tsconfig.cli.json' })],
   },
-  plugins: [
-    typescript({ tsconfig: './tsconfig.cli.json' }),
-    shebang(),
-    json(),
-    nodeResolve({ preferBuiltins: true }),
-    commonjs(),
-    cleanup(),
-  ],
-};
+  {
+    input: ['cli/upload/index.ts'],
+    output: {
+      dir: 'upload',
+      format: 'cjs',
+    },
+    plugins: [...commonPlugins, typescript({ tsconfig: './tsconfig.upload.json' })],
+  },
+];
